Once more refactoring! ImageHelper and Handler were deleted. I add some more repositories and now ImageConvertor boss can decide what type of convert to do.

+++ src/main/java/ru/shishlov/btf/repositories/ImageRepositoryFS.java
@@ -1,8 +1,7 @@
-package ru.shishlov.btf.components.images;
+package ru.shishlov.btf.repositories;
 
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Repository;
 import ru.shishlov.btf.entities.Image;
 
 import java.io.File;
@@ -10,42 +9,50 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.Optional;
 
 import static org.apache.tomcat.util.http.fileupload.FileUtils.deleteDirectory;
 
-
-@Component
-@PropertySource(value = "classpath:application.properties")
-public class ImageHandlerFS implements ImageHandler {
+@Repository
+public class ImageRepositoryFS implements ImageRepository{
+    private final ImageRepositoryJpa imageRepositoryJpa;
     @Value(value = "${images.path}")
     private String imagePath;
 
+    public ImageRepositoryFS(ImageRepositoryJpa imageRepositoryJpa){
+        this.imageRepositoryJpa = imageRepositoryJpa;
+    }
+
     @Override
-    public void prepareForSave(Image image, String login) {
+    public void save(Image image, String login) {
         Path p = Paths.get(imagePath+"/"+login+"/"+image.getName());
         try {
+            if(!Files.exists(p))
             Files.createDirectories(p.getParent());
             Files.write(p, image.getContent());
             image.setLocation(p.toString());
             image.setContent(null);
+            imageRepositoryJpa.save(image);
         } catch (IOException e) {
             e.printStackTrace();
         }
     }
 
     @Override
-    public void prepareForUpdate(Image image) {
+    public void update(Image image) {
         Path p = Paths.get(image.getLocation());
         try {
+            if(Files.exists(p))
             Files.write(p, image.getContent());
             image.setContent(null);
+            imageRepositoryJpa.save(image);
         } catch (IOException e) {
             e.printStackTrace();
         }
     }
 
     @Override
-    public void prepareForDelete(String login) {
+    public void clean(String login) {
         Path p = Paths.get(imagePath + login);
         File f = new File(p.toString());
         try {
@@ -53,5 +60,10 @@ public class ImageHandlerFS implements ImageHandler {
         } catch (IOException e) {
             e.printStackTrace();
         }
+    }
+
+    @Override
+    public Optional<Image> findById(long id){
+        return imageRepositoryJpa.findById(id);
     }
 }
